Refactor autopad in the onnx.py and paddlepaddle.py to relay/frontend/common.py
